ANOFA: Analyses of Frequency Data

ANOFA is a library to perform frequency data analyses. It is based on the G statistics (first developed by Fisher). This statistics is fully additive and can be decomposed in main effects and interaction effects, in simple effects in the decomposition of a significant interaction, in contrasts, etc. The present library performs these analyses and also can be used to plan statistical power for the analysis of frequency, obtain plots of the various effects, etc. It aims at replicating the most commonly-used ANOVA commands so that using this package should be easy.

The data supplied to an ANOFA can be in three formats: (i) long format, (ii) wide format, (iii) compiled format, or (iv) raw format. Check the anofa commands for more precision (in what follow, we assume the compiled format where the frequencies are given in a column name Freq)

The main function is

w <- anofa(formula, data)

where formula is a formula giving the factors, e.g., "Freq ~ A * B".

For more details on the underlying math, see Laurencelle and Cousineau (2023) .

An omnibus analysis may be followed by simple effects or contrasts analyses: emFrequencies(w, formula) contrast(w, listOfContrasts)

As usual, the output can be obtained with print(w) #implicite summary(w) # or summarize(w) for the G statistics table explain(w) # for human-readable output

Data format can be converted to other format with toLong(w) toWide(w) toCompiled(w) toRaw(w) toTabulated(w) # the only format that cannot be used as input to anofa

The package includes additional, helper, functions:

powerXXX() to compute sample size given effect size;

anofaPlot() to obtain a plot of the frequencies with error bars;

effectsizeXXX() to compute the effect size;

grf() to generate random frequencies from a given design.

and example datasets, some described in the article:

LandisBarrettGalvin2013 illustrates a 5 x 3 design;

LightMargolin1971 illustrates a 5 x 2 design;

Gillet1993 illustrates a 2 x 3 x 2 design;

Detergent illustrates a 2 x 2 x 2 x 3 design;

The functions uses the following options:

ANOFA.feedback ((currently unused));

ANOFA.digits for the number of digits displayed in G statistics tables.

Details

ANOFA library for analyses of frequency data

References

Laurencelle L, Cousineau D (2023). “Analysis of frequency tables: The ANOFA framework.” The Quantitative Methods for Psychology, 19, 173--193. doi:10.20982/tqmp.19.2.p173 .

See also

Author

Maintainer: Denis Cousineau denis.cousineau@uottawa.ca

Other contributors:

Louis Laurencelle louis.laurencelle@gmail.com [contributor]

Pier-Olivier Caron pocaron19@gmail.com [contributor]